requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-multiselectfield==0.1.12
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
greenlet==1.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
stripe==2.56.0
uritemplate==3.0.1
urllib3==1.26.4
webdriver-manager==3.4.1
